+++ zucchini-ui-backend/src/main/java/io/zucchiniui/backend/scenario/views/StepOccurrence.java
@@ -1,0 +1,23 @@
+package io.zucchiniui.backend.scenario.views;
+
+import io.zucchiniui.backend.scenario.domain.StepStatus;
+import io.zucchiniui.backend.shared.domain.BasicInfo;
+
+public class StepOccurrence {
+
+    private final StepStatus status;
+    private final BasicInfo info;
+
+    StepOccurrence(BasicInfo info, StepStatus status) {
+        this.info = info;
+        this.status = status;
+    }
+
+    public StepStatus getStatus() {
+        return status;
+    }
+
+    public BasicInfo getInfo() {
+        return info;
+    }
+}
