+++ zucchini-ui-frontend/src/styles/main.less
@@ -148,7 +148,7 @@
   opacity: 0;
 }
 
-.modal-dialog {
+.details-modal-dialog {
   width: 90%;
   display: table;
 }
@@ -173,4 +173,8 @@
 .step-definition-argument {
   font-weight: bold;
   color: #FF8D1E;
+}
+
+.step-definition-variant-details {
+  float: left;
 }
